engine: extract dispatch::combat (#159)
